add cucumber steps catalog refactor baraholka
